refactor: componentize search modal & create reusable wrapper
